requirements.txt
python-decouple==3.1
Django==1.11
djangorestframework==3.8.2
